requirements.txt
matplotlib==3.5.1
numpy==1.22.2
Pillow==9.2.0
scikit_learn==1.1.1
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers==4.19.2
